Feature/prime and big group keys

+++ src/test/resources/db/component/component_link_by_link_and_pos.sql
@@ -1,4 +1,4 @@
-PREPARE component_link_by_link_and_pos (int, int, int, int) AS
+PREPARE component_link_by_link_and_pos (bigint, bigint, bigint, bigint) AS
     SELECT     s.component_link_id,
                u.component_link_id                                                          AS user_component_link_id,
                s.user_id,
