+++ src/test/resources/db/component/component_link_by_link_ids.sql
@@ -1,4 +1,4 @@
-PREPARE component_link_by_link_ids (int, int, int) AS
+PREPARE component_link_by_link_ids (bigint, bigint, bigint) AS
     SELECT     s.component_link_id,
                u.component_link_id                                                          AS user_component_link_id,
                s.user_id,
