+++ src/test/resources/db/formula/formula_link_by_id.sql
@@ -1,4 +1,4 @@
-PREPARE formula_link_by_id (int, int) AS
+PREPARE formula_link_by_id (bigint, bigint) AS
     SELECT
                s.formula_link_id,
                u.formula_link_id AS user_formula_link_id,
