+++ src/test/resources/db/ref/source_list_by_name_like.sql
@@ -1,4 +1,4 @@
-PREPARE source_list_by_name_like (int,text) AS
+PREPARE source_list_by_name_like (bigint,text) AS
     SELECT s.source_id,
            u.source_id                                                                                 AS user_source_id,
            s.user_id,
