+++ src/test/resources/db/triple/triple_list_by_ids.sql
@@ -1,4 +1,4 @@
-PREPARE triple_list_by_ids (int, int[]) AS
+PREPARE triple_list_by_ids (bigint, bigint[]) AS
     SELECT s.triple_id,
            u.triple_id AS user_triple_id,
            s.user_id,
