+++ src/test/resources/db/triple/triple_list_by_phr_lst_up.sql
@@ -1,4 +1,4 @@
-PREPARE triple_list_by_phr_lst_up (int, int[]) AS
+PREPARE triple_list_by_phr_lst_up (bigint, bigint[]) AS
     SELECT s.triple_id,
            u.triple_id AS user_triple_id,
            s.user_id,
