+++ src/test/resources/db/word/word_list_by_type.sql
@@ -1,4 +1,4 @@
-PREPARE word_list_by_type (int, int) AS
+PREPARE word_list_by_type (bigint, bigint) AS
     SELECT
             s.word_id,
             u.word_id AS user_word_id,
