+++ src/test/resources/db/word/word_std_by_id.sql
@@ -1,4 +1,4 @@
-PREPARE word_std_by_id (int) AS
+PREPARE word_std_by_id (bigint) AS
     SELECT word_id,
            word_name,
            values,
